Updated playtime command to use longs all the way. Updated old database queries to use prepared statements over *string snakes*

+++ src/main/java/com/kuborros/FurBotNeo/commands/MusicCommands/MusicTimeCommand.java
@@ -60,7 +60,7 @@ public class MusicTimeCommand extends MusicCommand {
             } else {
                 val = val.trim();
             }
-            int seconds;
+            long seconds;
             try {
                 seconds = (min ? 60 : 1) * Integer.parseInt(val);
                 long milis = (seconds * 1000);
